# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi documentación de Minibares",
        "description": "Class LogController",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/log_not_founds": {
            "get": {
                "tags": [
                    "LogNotFound"
                ],
                "summary": "List all log_not_founds",
                "description": "Display a listing of the resource.",
                "operationId": "df5aa9190fb356f69b7dccaa21af1876",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/log_not_founds": {
            "post": {
                "tags": [
                    "LogNotFound"
                ],
                "summary": "Crear un log not found",
                "description": "Store a newly created resource in storage.",
                "operationId": "a6257d6261fce308253b67ecb537b95a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fechaCreacion": {
                                        "type": "string"
                                    },
                                    "usuarioId": {
                                        "type": "string"
                                    },
                                    "usuarioNombre": {
                                        "type": "string"
                                    },
                                    "usuarioRolesId": {
                                        "type": "string"
                                    },
                                    "nombreEmpresa": {
                                        "type": "string"
                                    },
                                    "controlador": {
                                        "type": "string"
                                    },
                                    "accion": {
                                        "type": "string"
                                    },
                                    "codigoError": {
                                        "type": "string"
                                    },
                                    "mensage": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "fechaCreacion": "2020-01-01T00:00:00",
                                    "usuarioId": "1",
                                    "usuarioNombre": "Admin",
                                    "usuarioRolesId": "1",
                                    "nombreEmpresa": "MyCompany",
                                    "controlador": "Inventarios",
                                    "accion": "index",
                                    "codigoError": "404",
                                    "mensage": "No existen inventarios"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crear un log not found."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/log_not_founds/{id}": {
            "get": {
                "tags": [
                    "LogNotFound"
                ],
                "summary": "Mostrar información de un log not found específico",
                "description": "Display the specified resource.",
                "operationId": "e162b944a15d8bb903ff674993860221",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar información de un log not found específico."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "LogNotFound"
                ],
                "summary": "Actualizar un log not found específico",
                "description": "Update the specified resource in storage.",
                "operationId": "02ac35a1756053f4204ef3181d1720d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fechaCreacion": {
                                        "type": "string"
                                    },
                                    "usuarioId": {
                                        "type": "string"
                                    },
                                    "usuarioNombre": {
                                        "type": "string"
                                    },
                                    "usuarioRolesId": {
                                        "type": "string"
                                    },
                                    "nombreEmpresa": {
                                        "type": "string"
                                    },
                                    "controlador": {
                                        "type": "string"
                                    },
                                    "accion": {
                                        "type": "string"
                                    },
                                    "codigoError": {
                                        "type": "string"
                                    },
                                    "mensage": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "fechaCreacion": "2020-01-01T00:00:00",
                                    "usuarioId": "1",
                                    "usuarioNombre": "Admin",
                                    "usuarioRolesId": "1",
                                    "nombreEmpresa": "MyCompany",
                                    "controlador": "Inventarios",
                                    "accion": "index",
                                    "codigoError": "404",
                                    "mensage": "No existen inventarios"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizar un log not found específico."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "LogNotFound"
                ],
                "summary": "Eliminar un log not found específico",
                "operationId": "09569db5550d1b0a49f7f7ca357bb343",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminar un log not found específico."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}